조건 검색 api 완성
