+++ back/src/main/java/teamparkinglot/parkinggo/history/repository/HistoryRepository.java
@@ -1,8 +1,9 @@
-package teamparkinglot.parkinggo.history;
+package teamparkinglot.parkinggo.history.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import teamparkinglot.parkinggo.history.entity.History;
 
 import java.util.List;
 import java.util.Optional;
