+++ back/src/main/java/teamparkinglot/parkinggo/history/repository/HistoryRepositoryQueryDsl.java
@@ -1,14 +1,15 @@
-package teamparkinglot.parkinggo.history;
+package teamparkinglot.parkinggo.history.repository;
 
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.stereotype.Repository;
+import teamparkinglot.parkinggo.history.entity.QHistory;
 import teamparkinglot.parkinggo.parking.dto.ParkingRecentDto;
 
 import javax.persistence.EntityManager;
 import java.util.List;
 
-import static teamparkinglot.parkinggo.history.QHistory.*;
+import static teamparkinglot.parkinggo.history.entity.QHistory.*;
 
 @Repository
 public class HistoryRepositoryQueryDsl {
@@ -22,7 +23,7 @@ public class HistoryRepositoryQueryDsl {
     public List<ParkingRecentDto> findRecentSearch(String email) {
         return queryFactory
                 .select(Projections.fields(ParkingRecentDto.class,
-                        history.parking.parkingName.as("parkingName"), history.parking.address.parcel.as("address")))
+                       history.parking.id.as("parkingId") , history.parking.parkingName.as("parkingName"), history.parking.address.parcel.as("address")))
                 .from(history)
                 .where(history.member.email.eq(email))
                 .orderBy(history.id.desc())
