+++ back/src/main/java/teamparkinglot/parkinggo/member/controller/MemberController.java
@@ -9,7 +9,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import teamparkinglot.parkinggo.exception.BusinessException;
 import teamparkinglot.parkinggo.exception.ExceptionCode;
-import teamparkinglot.parkinggo.mail.MailService;
+import teamparkinglot.parkinggo.member.mail.MailService;
 import teamparkinglot.parkinggo.member.dto.*;
 import teamparkinglot.parkinggo.member.entity.Member;
 import teamparkinglot.parkinggo.member.mapper.MemberMapper;
@@ -17,8 +17,8 @@ import teamparkinglot.parkinggo.member.service.MemberService;
 import teamparkinglot.parkinggo.reservation.dto.ReservationResponseDto;
 import teamparkinglot.parkinggo.reservation.service.ReservationService;
 import teamparkinglot.parkinggo.security.principal.PrincipalDetails;
-import teamparkinglot.parkinggo.uuid.UuidService;
-import teamparkinglot.parkinggo.uuid.Uuid;
+import teamparkinglot.parkinggo.member.uuid.UuidService;
+import teamparkinglot.parkinggo.member.uuid.Uuid;
 
 import javax.mail.MessagingException;
 import javax.validation.Valid;
