+++ back/src/main/java/teamparkinglot/parkinggo/member/mail/MailService.java
@@ -1,4 +1,4 @@
-package teamparkinglot.parkinggo.mail;
+package teamparkinglot.parkinggo.member.mail;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
