+++ back/src/main/java/teamparkinglot/parkinggo/member/uuid/UuidService.java
@@ -1,4 +1,4 @@
-package teamparkinglot.parkinggo.uuid;
+package teamparkinglot.parkinggo.member.uuid;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
